# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Clubs SaaS API",
        "description": "API documentation for Clubs SaaS",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/Clubs/public/api",
            "description": "Local Server"
        }
    ],
    "paths": {
...
        "/v1/notification-logs": {
            "get": {
                "tags": [
                    "Notification Management"
                ],
                "summary": "📜 View notification history",
                "operationId": "d72e0f2f9d91c446c4eec0013cdd3a87",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/v1/notification-logs/{id}": {
            "get": {
                "tags": [
                    "Notification Management"
                ],
                "summary": "🔍 View single log detail",
                "operationId": "5342553fcc980d241790c5e9ab869a62",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Notification Management"
                ],
                "summary": "🗑 Delete a log entry",
                "operationId": "6ded746a8e12806c20e4a8ba2bc02df7",
                "responses": {
                    "200": {
                        "description": "Log deleted"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/v1/notification-stats": {
            "get": {
                "tags": [
                    "Notification Management"
                ],
                "summary": "📊 Notification delivery statistics",
                "operationId": "32d5544daf0d3717cb2f4936063dfd26",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/v1/notification-templates": {
            "get": {
                "tags": [
                    "Notification Management"
                ],
                "summary": "📋 List notification templates",
                "operationId": "4dcfe778ed07a46185452dddc88bef96",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Notification Management"
                ],
                "summary": "➕ Create template",
                "operationId": "9b988eb8fe57b6f6ad112e5a6c3f1244",
                "responses": {
                    "201": {
                        "description": "Template created"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/v1/notification-templates/{id}": {
            "get": {
                "tags": [
                    "Notification Management"
                ],
                "summary": "🔍 View template details",
                "operationId": "f5430ebf7b74e3bd4a421f3b32fddfb7",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Notification Management"
                ],
                "summary": "Edit a template",
                "operationId": "9e29121b7c0097105646fe7a6d6507eb",
                "responses": {
                    "200": {
                        "description": "Template updated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Notification Management"
                ],
                "summary": "🗑 Delete a template",
                "operationId": "b8c13243ea54d8f12ff0ddfa8fecbf2f",
                "responses": {
                    "200": {
                        "description": "Template deleted"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/v1/notification-templates/{id}/toggle": {
            "post": {
                "tags": [
                    "Notification Management"
                ],
                "summary": "🔘 Enable or disable a template",
                "operationId": "264cfcb9cb6de83ab8eb372f9fa87b59",
                "responses": {
                    "200": {
                        "description": "Status toggled"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/v1/notification-templates/{slug}/test": {
            "post": {
                "tags": [
                    "Notification Management"
                ],
                "summary": "🧪 Send a test notification",
                "operationId": "c589672df25790303d66e83849aa048a",
                "responses": {
                    "200": {
                        "description": "Test sent"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/v1/staff": {
            "get": {
                "tags": [
                    "Staff Management"
                ],
                "summary": "👥 List all staff and coaches",
                "operationId": "38db582bc7fab639cc88c4c4f9df2c2c",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Staff Management"
                ],
                "summary": "➕ Onboard a new staff member",
                "operationId": "d9c544ebcfbbc6077d35711e4060448d",
                "responses": {
                    "201": {
                        "description": "Staff onboarded"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/v1/staff/{id}/schedule": {
            "post": {
                "tags": [
                    "Staff Management"
                ],
                "summary": "📅 Set staff weekly schedule",
                "operationId": "447bba7e0c60e3c4aba9352ef908da95",
                "responses": {
                    "200": {
                        "description": "Schedule updated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/v1/staff/{id}/check-in": {
            "post": {
                "tags": [
                    "Staff Management"
                ],
                "summary": "🕒 Staff check-in",
                "operationId": "6fd74244a3e90fbb1bae00d92e5e8468",
                "responses": {
                    "200": {
                        "description": "Check-in successful"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}